Develop/1.0

+++ src/pages/MyPage/DetailForm/MakeFormList/index.tsx
@@ -2,19 +2,19 @@ import { useRecoilValue } from 'recoil';
 import Button from '../../../../components/ui/Button';
 import { BottomBox, ButtonWrapper, FormListWrapper, HeaderWrapper, Title } from '../styles';
 import { color } from '../../../../recoil/Color/atom';
-import { useCallback, useState } from 'react';
-import { useInfiniteQuery, useMutation, useQuery } from 'react-query';
-import { deleteFrom, getMakeForms, getPagingInfo } from '../../../../api/getFormInfo';
+import { useCallback, useRef } from 'react';
+import { useInfiniteQuery, useMutation, useQueryClient } from 'react-query';
+import { deleteFrom, getPagingInfo } from '../../../../api/getFormInfo';
 import { Col, Row } from 'antd';
-import { GetForm } from '../../../../typings/getForm';
 import { useNavigate } from 'react-router-dom';
 import InfiniteScroll from 'react-infinite-scroller';
-import { formInfo } from '../../../../recoil/MakeForm/atom';
+import { useMessage } from '../../../../hooks/useMessage';
 
 export default function MakeFormList() {
   const { blue, lightPurple } = useRecoilValue(color);
   const navigate = useNavigate();
-  const page = 0;
+  const { showMessage, contextHolder } = useMessage();
+  const ref = useRef<number | null>(null);
 
   const {
     data,
@@ -24,17 +24,22 @@ export default function MakeFormList() {
     isFetching: infiniteIsFetching,
     isError: infiniteIsError,
     error,
-  } = useInfiniteQuery('makeForms', ({ pageParam = page }) => getPagingInfo({ userId: 1, pageParam }), {
-    getNextPageParam: (lastPage) => {
-      if (lastPage?.nextPage) {
-        // return page++;
-      } else return undefined;
-    },
-  });
+  } = useInfiniteQuery(
+    'makeForms',
+    ({ pageParam = 0 }) => getPagingInfo({ userId: 1, pageParam }), // pageParam의 초기값을 0으로 설정
+    {
+      getNextPageParam: (lastPage, allPages) => {
+        if (lastPage?.nextPage) {
+          return allPages.length + 1;
+        }
+      },
+      refetchOnMount: false, // 마운트(리렌더링)될 때 데이터를 다시 가져오지 않음
+      refetchOnWindowFocus: false, // 브라우저를 포커싱했을때 데이터를 가져오지 않음
+      refetchOnReconnect: false, // 네트워크가 다시 연결되었을때 다시 가져오지 않음
+    }
+  );
 
-  console.log('여기', data?.pages[0]?.pagingData);
-
-  // const { data: makeFormInfo, isLoading, error, isError } = useQuery<GetForm[]>('myMakeForm', getMakeForms);
+  const queryClient = useQueryClient();
 
   const {
     mutate: deleteMutate,
@@ -42,15 +47,36 @@ export default function MakeFormList() {
     isError: deleteIsError,
     error: deleteError,
     isSuccess: deleteIsSuccess,
-  } = useMutation(deleteFrom);
+  } = useMutation(deleteFrom, {
+    onMutate: async (id: number) => {
+      const snapshot = queryClient.getQueryData('makeForms');
+
+      queryClient.setQueryData('makeForms', (old: any) =>
+        old.pages[ref.current!].pagingData.filter((item: any) => {
+          return item.id !== id;
+        })
+      );
+
+      return { snapshot };
+    },
+    onError: (error, newData, context) => {
+      if (context?.snapshot) {
+        queryClient.setQueryData('makeForms', context.snapshot);
+        showMessage('error', '삭제에 실패했습니다.');
+      }
+    },
+    onSettled() {
+      queryClient.invalidateQueries('makeForms');
+    },
+  });
 
   const deleteForm = useCallback((id: number) => {
-    deleteMutate(id);
+    const confirmDelete = window.confirm('이 설문을 삭제하시겠습니까?');
+    if (confirmDelete) {
+      deleteMutate(id);
+      showMessage('success', '삭제되었습니다.');
+    }
   }, []);
-
-  const loadMoreItems = useCallback(() => {
-    fetchNextPage();
-  }, [fetchNextPage]);
 
   if (infiniteIsLoading) return <div>loading...</div>;
   if (infiniteIsError) return <div>error...</div>;
@@ -64,95 +90,60 @@ export default function MakeFormList() {
         </HeaderWrapper>
 
         {infiniteIsFetching && <div>loading...</div>}
-        <InfiniteScroll loadMore={loadMoreItems} hasMore={hasNextPage}>
+        <InfiniteScroll loadMore={() => fetchNextPage({ pageParam: data?.pages.length })} hasMore={hasNextPage}>
           <FormListWrapper>
-            {data?.pages[0]?.pagingData?.map((formInfo) => (
-              <div key={formInfo.id}>
-                <Title>
-                  <span>{formInfo.title}</span>
-                  <span>{formInfo.content}</span>
-                </Title>
-                <ButtonWrapper>
-                  <div>
-                    <Button
-                      onClick={() => navigate(`/mypage/editform/${formInfo.id}`)}
-                      color={'black'}
-                      bgColor={lightPurple}
-                      fontSize={1.3}
-                      width={4}
-                      height={8}
-                    >
-                      수정
-                    </Button>
-                    <Button
-                      onClick={() => deleteForm(formInfo.id)}
-                      color={'black'}
-                      bgColor={lightPurple}
-                      fontSize={1.3}
-                      width={4}
-                      height={8}
-                    >
-                      삭제
-                    </Button>
-                  </div>
-                  <div>
-                    <Button
-                      onClick={() => navigate(`/statistic/${formInfo.id}`)}
-                      color={'black'}
-                      bgColor={blue}
-                      fontSize={1.3}
-                      width={10}
-                      height={3.5}
-                    >
-                      응답보기
-                    </Button>
-                  </div>
-                </ButtonWrapper>
-              </div>
-            ))}
+            {contextHolder}
+            {data?.pages?.map((page) =>
+              page?.pagingData?.map((formInfo, idx) => (
+                <div key={formInfo.id}>
+                  <Title>
+                    <span>{formInfo.title}</span>
+                    <span>{formInfo.content}</span>
+                  </Title>
+                  <ButtonWrapper>
+                    <div>
+                      <Button
+                        onClick={() => navigate(`/mypage/editform/${formInfo.id}`)}
+                        color={'black'}
+                        bgColor={lightPurple}
+                        fontSize={1.3}
+                        width={4}
+                        height={8}
+                      >
+                        수정
+                      </Button>
+                      <Button
+                        onClick={() => {
+                          deleteForm(formInfo.id);
+                          ref.current = idx;
+                        }}
+                        color={'black'}
+                        bgColor={lightPurple}
+                        fontSize={1.3}
+                        width={4}
+                        height={8}
+                      >
+                        삭제
+                      </Button>
+                    </div>
+                    <div>
+                      <Button
+                        onClick={() => navigate(`/statistic/${formInfo.id}`)}
+                        color={'black'}
+                        bgColor={blue}
+                        fontSize={1.3}
+                        width={10}
+                        height={3.5}
+                      >
+                        응답보기
+                      </Button>
+                    </div>
+                  </ButtonWrapper>
+                </div>
+              ))
+            )}
           </FormListWrapper>
         </InfiniteScroll>
-
-        {/* {makeFormInfo?.map((formInfo) => { */}
-        {/*   return ( */}
-        {/*     <div key={formInfo.id}> */}
-        {/*       <Title> */}
-        {/*         <span>{formInfo.title}</span> */}
-        {/*         <span>{formInfo.content}</span> */}
-        {/*       </Title> */}
-
-        {/*       <ButtonWrapper> */}
-        {/*         <div> */}
-        {/*           <Button */}
-        {/*             onClick={() => navigate(`/mypage/editform/${formInfo.id}`)} */}
-        {/*             color={'black'} */}
-        {/*             bgColor={lightPurple} */}
-        {/*             fontSize={1.3} */}
-        {/*             width={4} */}
-        {/*             height={8} */}
-        {/*           > */}
-        {/*             수정 */}
-        {/*           </Button> */}
-        {/*           <Button */}
-        {/*             onClick={() => deleteForm(formInfo.id)} */}
-        {/*             color={'black'} */}
-        {/*             bgColor={lightPurple} */}
-        {/*             fontSize={1.3} */}
-        {/*             width={4} */}
-        {/*             height={8} */}
-        {/*           > */}
-        {/*             삭제 */}
-        {/*           </Button> */}
-        {/*         </div> */}
-        {/*         <div> */}
-        {/*           <Button color={'black'} bgColor={blue} fontSize={1.3} width={10} height={3.5}> */}
-        {/*             응답보기 */}
-        {/*           </Button> */}
-        {/*         </div> */}
-        {/*       </ButtonWrapper> */}
-        {/*     </div> */}
-        {/*   ); */}
-        {/* })} */}
         <BottomBox />
       </Col>
